+++ v3io/controlplane/exceptions.py
@@ -1,10 +1,3 @@
-class RetryUntilSuccessfulInProgressErrorMessage(Exception):
-    def __init__(self, message, *, variables=None):
-        super().__init__(message)
-        self.variables = variables if variables else {}
-        self.message = message
-
-
 class ResourceDeleteException(Exception):
     def __str__(self):
         return "This resource is not delete-able"
